Move creatable actionable types to /models
