+++ src/models/index.ts
@@ -35,3 +35,4 @@ export * from './timeframe';
 export * from './user';
 export * from './user-preferences';
 export * from './version';
+export * from './_actionable';
